[FIX] industry_fsm: mark as done button works for 3 step delivery for fsm_tasks
This commit's purpose is to fix the mark as done button for fsm task. It launch a traceback
when trying to validate more than one delivery at the time (e.a. in multi
step route) if the report setting is enabled for the inventory app.
It works fine now

community PR : https://github.com/odoo/odoo/pull/113628
task-3204596

closes odoo/enterprise#40552

X-original-commit: b9b78248029b8b2fdf341f0d2d6c92af39d5ebde
Related: odoo/odoo#120159
Signed-off-by: Tiffany Chang <tic@odoo.com>