[IMP] web_enterprise: PWA "offline" mode
This commit introduces a way to notify the user that he's "offline"
(aka. cannot reach it's Odoo server) and that Odoo doesn't work in a
graceful way in this circumstance.

To do so, the Service-Worker will return the response of the
´web/offline´ route, which is cached at its setup.

Note: this screen is only show when launched while "offline" and fails
to load the requested page. It does not "interrupt" the WebClient to
show this screen when the connection drops off (cf. not a replacement
for the existing notification).

Task ID: 3203639

closes odoo/enterprise#40489

Signed-off-by: Pierre Paridans (app) <app@odoo.com>