[FIX] mrp_mps: use correct id in schedule stat button
Before this commit, the context added to the schedule stat button
contained a reference to the active_id value. However, in this case, we
actually want the current id of the record that we are looking at, not
the id from a previous view.

Note that as far as I can tell, almost every use of `active_id` in stat
button is wrong. It just usually works because we don't have an active
id set, and there is a fallback in the basicmodel to get the current
record id instead.  However, if a user would get to that form view by
following some other relations and coming from another form view, then
the active_id would not be correct.

closes odoo/enterprise#40514

X-original-commit: 241dc58497ba78e83ed93e4c1b7d9974b99d3292
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>