[FIX] spreadsheet_edition, web_cohort: fix untranslatable text
This is due to Babel—the library we use to extract `_t` occurrences from
the sources—ignoring javascript template literals, resulting in any
gettext inside template literals not being parsed and exported as
translatable resources.

From version 2.12.0 of the library, support for javascript template
literals can be enabled by providing the `parse_template_string` option
to the Babel `extract` method.

Unfortunately, the library cannot be updated for now, as our policy is
to stick to the version of packages from the official Debian repository.

This commit refactors some code so that no gettext is inlined within
template literals.

closes odoo/enterprise#40457

Related: odoo/odoo#119996
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>