[FIX] website_sale_renting: fix date consistency issues
Currently, when a customer clicks on a renting product name, while in
the cart, he is redirected to the product page. The date picker should
show the date of the sale order's first sale order line but it fails and
shows the date in the wrong timezone because the timezone between the
cart and the product page is not well handled.

Now, the date picker shows the date of the sale order's first sale order
line.

Also: drop moment.js in favor of luxon

task-3293848

closes odoo/enterprise#40534

Signed-off-by: Valentin Chevalier <vcr@odoo.com>