[FIX] (website_helpdesk_)knowledge: saving external inserts
This commit fixes a bug where external insertion inside an article would not
be saved if you clicked on a record inside the view or opened the link.
This is caused because the article was not flagged as dirty when opening either
a record or a link which does not trigger a save by the field html.

Now when we are finished inserting we are asking the record to check for
changes which will flag the field as dirty and will trigger the save mechanism
on opening an external record/link.

We also adapted one of the tour in website_helpdesk_knowledge that test cross
module features or knowledge to take this behavior into account.

task-3292309

closes odoo/enterprise#40486

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>