[IMP] web_enterprise: popover background
This commits is related to the property modal redesign. In order to reach
more consistency in our variables use, we defined that the
`$popover-border-color` should match the default `$border-color` variable.

The issue we are facing is the fact that `$border-color` uses the BS's
`$o-gray-300` variable in dark mode, which is the same value as our
popover background. Opening a popover in dark mode would then make these
borders invisible which is not the expected behavior.

What we do here is to modify the popover background in dark mode to make
sure borders are visible.

We also introduce a modification of the JS test related to the knowledge
properties. Before this commit, the structure of the modal was composed
of a header and a body.

Since we refactored the component in one single block with the class
`.o_modal_container`, we apply the `.o_field_property_definition_header`
class directly on the input rather than targeting an input inside the
header.

task-3084756

closes odoo/enterprise#43091

X-original-commit: 1cf7d7aff069d4066af7a2ab4e6423757f1ceefa
Related: odoo/odoo#126274
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>