[FIX] account_reports: fix closing vat options
Some new options are generated when constructing the query for getting
the lines of the closing entry. These options make use of the fiscal
period boundry dates of the current company, and not the dates of the of
the current options.

However, the current code fails to recompute the options based on these
new dates, and subsequently the column group keys are not re-generated
to utilise the correct dates.

This commit changes these options to utilise a 'custom' date and a
'custom' filter, and recomputes them based on this new date information.

closes odoo/enterprise#20245

Related: odoo/odoo#75634
Signed-off-by: Josse Colpaert <jco@odoo.com>