[FIX] delivery_ups: round unit value before sending to UPS
Before this commit, in case of a rounding error in the calculation of
monetary value of a commodity, we get an error from UPS:
`Invalid or missing product unit value for product number 1.
Valid length is 1 to 19 numeric and it can hold up to 6 decimal places`

To avoid this, we need to round the unit price before sending it to the
UPS. We are rounding to 2 decimal places because there is not a
straight-forward way to get the number of decimal places of the
currency and it covers most cases.

opw-3289042

closes odoo/enterprise#40536

Signed-off-by: Tiffany Chang <tic@odoo.com>