[FIX] planning: fix preview of shift re-assigned mail template
When using the Planning: Shift Re-Assigned email template, an error occurs when
clicking on the Preview button. This is due to the template trying to retrieve
data from the context, which may not always be present while previewing the mail,
resulting in a traceback error.

In this commit, the employee and assignee names are passed through the context
instead of records.

task-3304959

closes odoo/enterprise#42967

X-original-commit: 6611cdd2a4d72d636111974c9fd08fe2b30a626e
Related: odoo/odoo#126004
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>