[FIX] hr_contract_salary: make required personal info required in the salary configurator
Issue:
'document' type of personal informations were not required on the salary
configurator, even when they were set as required on the personal info.

Moreover, we set personal information as required, if it is set as
'Requested Document' for an advantage.

This changes also implied populating 'document' type personal informations
for the tours.

task - 3266175

closes odoo/enterprise#42619

Related: odoo/upgrade#4783
Signed-off-by: Kevin Baptiste <kba@odoo.com>