[FIX] web: update PayslipLineField and WorkedDaysField
Since the commit ac7236167a71954210b75f88b958c812e7399e80, the menu
action "Edit Payslip Lines" doesn't update all the payslip lines when one
of them is modified.

Why not ?
To perform an update in a Field, you must no longer use props.update
which only takes a value but props.record.update which takes a dictionary
containing the fields to be modified (key = fieldname).

Solution:
We must therefore adapt the overwrite in PayslipLineField and WorkedDaysField.
They should no longer look in the props to see which field is being edited,
but if the amount or quantity field is present in the changes pass it to update.

How to reproduce ?
- Go to the form view of a Payslip
- Click on the action menu "Edit Payslip Lines
- Edit the amount of a line

Before this commit:
    Only the edited line has changed. No save takes place.

After this commit:
    The edit line has been changed. The record is saved and the dependent
    lines have been modified.

closes odoo/enterprise#40619

X-original-commit: f6fdea8262737e014ac7a94b346634434c9552c3
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Georis François (fge) <fge@odoo.com>