[FIX] account_bank_statement_import: only display errors after all correct imports have been done
Currently, when importing multiple bank statements, if one of them has an error, we block the importation of the others.

With this commit, we will process all the statements and, if applicable, show an error message at the end. This allows correct statements to be imported even if one in the batch is failing.

Therefore, we now have three flows:
- Everything went smoothly, we can redirect the user to correct view
- All the imports failed, we show an error
- At least one import succeeded and at least one failed, we display a warning message containing the names of the attachments that were not imported and show an option to view the ones that were correctly imported

task id 2747359

closes odoo/enterprise#40519

X-original-commit: ea10325
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>