[FIX] appointment: custom meeting slot
This commit fixes the case where an operator's calendar would contain meetings
where he is still available (using the "show_as" field).

Currently the appointment availability checks block all slots that conflict
with a calendar.event, regardless of the "show_as" flag.

We fix it by accommodating the necessary domains.

A test was added to ensure this behavior.

Task-3264505

closes odoo/enterprise#43269

X-original-commit: 32f9e4ce2c817b51ca1630caad23a6d21d769092
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>