[FIX] knowledge: keep optional fields in fullscreen mode in embeds
This commit fixes a bug where you would lose the optional columns set inside
the embedded view when opening the view in fullscreen.
This bug occurs because the optional columns are stored inside the localStorage
under a key generated by the renderers, but in Knowledge we patched each view
type to add the id of the embedded view, in order to have embedded views with
different configuration.

The problem is that when opening the embed in fullscreen the embedId is no
longer taken into account inside the optionalFieldsKey, meaning that the view
goes to its default state.
We changed the override of `createKeyOptionalFields` inside
`views_renderers_patches` to take into account the embeddedViewId that we are
now adding in the additionalContext when executing the action of opening the
view in fullscreen.

task-3293109

Part-of: odoo/enterprise#40635