[FIX] knowledge: removes members banner when internal_permission='none'
This commit fixes a bug where the banner for member_only article would appear
in private articles as a user that cannot access it.
It was appearing because the invisible modifier was not complete, we needed to
also make it invisible when the internal_permission field was set to 'none',
meaning when the article is only accessible to the members.

Now we added this check in the invisible modifier for the banner and it does
not appear when the user has no permission on the opened article.

task-3293200

closes odoo/enterprise#40742

X-original-commit: 724445c97172ad3fead956c72408c7e641b41951
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>