[IMP] timesheet_grid: add default filter on timer fields
Before this commit searching on project and task display all
records in search more dialog without distinction and because
of that sometimes it is difficult for user to find a proper
record.

This commit add default filter on search more dialog to
display user related record.

task-3251672

closes odoo/enterprise#39936

Related: odoo/odoo#118911
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>