[IMP] knowledge: allow embedded view for gantt view type
With this commit, users will be able to embed the Gantt view directly
within the knowledge editor, directly in the body of articles, and it
also allows linking the Gantt view.

To insert a Gantt view, users have to:

Open the Gantt view;
Click on the 'Favorite' dropdown menu under the search bar of the
control panel;
Click on the 'Insert view in article' menu item for inserting the view;
and;
Click on the 'Insert link in article' menu item for inserting the link.

Task-3297331

closes odoo/enterprise#40664

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>