[IMP] (website_)knowledge: add portal link inside the `my` page
This commit adds an entry inside the `my` homepage for portal users, this entry
shows the number of articles that portal user has access to and redirects the
user to Knowledge's frontend view.

We are adding a new template inside the views folder that inherits from the
portal template for the user's `my` page and adds the entry after the others.
We are also adding a portal controller that redirects the user to their
Knowledge form view using a redirect to the route `/knowledge/home` so that we
use route that already exists for future portal enhancements.

task-3299238

closes odoo/enterprise#40659

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>