[FIX] project: fix sub-tasks notebook shadow issue
Description of the issue/feature this PR addresses:
In the website when task of project is open, sub tasks notebook has dark shadow
below the list. Shadow should be same as sub tasks list in backend.

Current behavior before PR:
Sub tasks notebook has dark shadow below the list in website.

Desired behavior after PR is merged:
Sub tasks notebook has same shadow as in backend project module.

causes:
in project sharing, gradient of 0.4 opacity is apply on shadow for dark mode in
list_controller.dark.scss file so that shadow effect is shown in dark mode
which is also apply on light mode therefore shadow is dark in light mode

Fix:
As all the list view in project sharing have same issue, so we remove
list_controller.dark.scss file in manifest project.webclient, so that this
effect is not apply in frontend.

task-3217427

closes odoo/enterprise#44171

X-original-commit: 82f9dc4934d459e4ddc8404e0e6af48ebe0139ce
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>