[FIX] planning: fix _compute_planning_hours_planned
Before this commit:
- sale_planning module:
* When a planning slot doesn't have a resource (resource_id = False),
its allocated_hours is not counted when computing planning_hours_planned
of its sale_line_id.
* The value of planning_hours_planned is wrong after installing
the demo data, the allocated hours of the last added planning_slot
are missing in the sum.

- project_timesheet_forecast_sale module:
When _compute_planning_hours_planned was inherited, the check on the material
resource was not taken into account

After this commit:
- sale_planning module:
* the bug is fixed,
technically: the where part of the query was changed from
```
WHERE resource_id" in
	(
		SELECT id
		FROM resource_resource
		WHERE resource_type != 'material'
		OR "resource_resource"."resource_type" IS NULL
	)
```
to
```
WHERE resource_id IS NULL
OR resource_id" in
        (
                SELECT id
                FROM resource_resource
                WHERE resource_type != 'material'
                OR "resource_resource"."resource_type" IS NULL
        )
```
*  Technically, the problem is when _compute_allocated_hours is called for
the planning slots linked to the concerned sale_line_id, the compute method
that should be called to compute planning_hours_planned after each modification
in the value of the allocated_hours field is not called so we have a wrong value.
So, in this commit, an add_to_compute is added to force that compute.

- project_timesheet_forecast_sale:
* The same previous condition was added to the query to filter by material
resources.

task-3299506

closes odoo/enterprise#40667

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>