[FIX] documents_project: fix issue when uploading file to any project
Before this commit:
When we go to the 'documents' app and try to upload a file to any
child workspace inside the 'Projects' workspace, it gives an
'Internal Server Error'.

Reason:
We are passing recordset of 'documents.folder' as a key to sort instead of id.

After this commit:
Now the files are uploaded without any error.

task-3284108

closes odoo/enterprise#42083

X-original-commit: 22e8b585a3dcd8cfc6a4854f9dc90e125ddff2c8
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>