[FIX] (website_)knowledge: use correct section when loading more roots
Purpose:
--------
Currently, the "load more" button uses the category of the section it is
in to only load articles of this category. However, in the portal view,
the "shared with you" section has articles from the shared category and
from the workspace category. Clicking on the load more button will only
load articles from one category (workspace articles), and some articles
will thus never be shown in the sidebar.
A similar issue arises in website_knowledge with the "public articles"
section. Clicking on the load more button will load workspace articles,
no matter if these articles are actually public (website_published) or
if the user can access them because he has been invited on these.

This commit modifies the domains used when loading more root articles in
the portal view, to load articles based on the section properties, not
on the section "category" (which is not relevant in portal view).

Task-3300677

Part-of: odoo/enterprise#40703