[FIX] website_knowledge: allow public users to load more articles
Purpose:
--------
In website_knowledge, the portal view can be accessed as a public user.
However, the route used to load more articles in the sidebar requires
users to  be logged in. Therefore, clicking on the load more button when
not logged does not work.

This commit changes the `auth` parameter of the load more route so that
public users can also use the load more button.

Task-3300677

closes odoo/enterprise#40703

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>