[IMP] currency_rate_live: bank of brazil API
Adds the central bank of brazil exchange rates service to the list of services.

As on some days the bank will return no result as there were no currency changes that day (mostly weekends and holidays), the date we send to the request gets decremented by one day until we find a value. While the API currently used fetches the changes on a specific date, there is another API that fetches the changes for a specific period (week for example). This API was not used instead as it was agreed that having a period of one week was too random and that we wouldn't be looping that much since the chance of going too many days backwards without a result is not that high, and in case this period had no result, the code will break.

Another issue was that there was no guarantee that all currencies are updated on a given date, which means instead of doing the backward loop on a date once, we'd have to do it for all currencies. To test this, the data in a period of six months was fetched for multiple currencies. It was found that for USD, EUR, COP and GBP: for every date any of those currencies had a result, all the rest had a result on the same date as well. Therefore, it was agreed, even though there is no guarantee, that it's not needed to do the backward loop on a date for every currency.

task-3270798

closes odoo/enterprise#40709

Signed-off-by: Quentin De Paoli <qdp@odoo.com>