[REF] web_studio: XMLResource editor better usage
This commit moves the XML editor for the ViewEditor a level above to prepare for new usages (ReportEditor).
It also adapts it to take into account the ace's edit session feature.

Lastly, this commit introduces a way to have a better display and understanding of the edited resources.
Namely, the main resource, in this case a view, is displayed with all the related resources and the logical
link between them. When a view has been modified by another via an inherit, it is visible, when a view t-calls
another, it is also visible in its own right.

This commit largely prepares the field for the ReportEditor

closes odoo/enterprise#43441

Related: odoo/odoo#126963
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>