[FIX] l10n_be_reports: fix partner VAT listing default date
Currently, the partner VAT listing uses the default computed
opening date, which is "Last Month".

But in practice, the date range that will almost always be
used is "Last Year".

This can be frustrating, especially on large databases where
the report will take a few seconds to load, only to be able
to select the correct date and wait some more to have the
desired result.

This is why the default opening date will now be set to
"Last Year" by default.

closes odoo/enterprise#40784

X-original-commit: 5dca4c1fb6ce09c09d8d09db63fe38eba478b362
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>