[IMP] web_studio: approval: implement server-side validation
Before, the approval mechanism validation was mostly implemented in the
frontend:

- The frontend detects that a button has a approval validation with
  `get_spec_approval`
- Whenever the button is pressed, the frontend contacted the server to
  know if the user can indeed call the method (with `check_approval`)
- If the server answered positively then the frontend did the call to
  the server to execute the method

The problem is that a tech-savy user could bypass this approval check:
- by manually calling the method on the model
- by calling a button doing the same thing outside a form view

Now, for method approval (in contrast to action approval) we do it in
the backend:
- The frontend detects that a button has a approval validation with
  `get_spec_approval`
- The frontend does not need to know if it is approved or not, it
  directly call the method on the model
- When receiving the call, the backend will now check if it is approved
  before executing it

Technically, this PR is inspired by `base.automation`. When an approval
rule is created/written then the corresponding method is patched to
verify approval rules.

If the environment is sudoed, approval check is entirely bypassed:
Several high-level features of Odoo call classic business methods with
an elevated (or sudoed) environment - e.g. when an order is placed in
the eCommerce, the public (or portal) user will call action_confirm on
the SO - whilst it could be seen as 'legitimate' to block such flows
using approvals, the truth of the matter is that there is no way for
users who set up approval flows to know what kind of flow they may be
affecting. Assuming that a user is able to anticipate that enabling
approvals on the 'Confirm' button of quotations will prevent eCommerce
orders from being validated is unreasonable.

task-id:3285734

closes odoo/enterprise#40800

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Co-authored-by: Damien Bouvy <dbo@odoo.com>