[FIX] knowledge: more accurate selector for kanban command tour
When removing the step delay from the test tours of the kanban cards,
the selector selecting the card to quickly create a record in the "Ongoing"
column was matching with the card that was automatically added by the
system after creating a record in the "New" column (see previous steps).

This commit will provide a more accurate selector to ensure that it
matches with the card of the "Ongoing" column and not the card that is
automatically added after creating a new record.

task-2867059

closes odoo/enterprise#29528

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>