[FIX] l10n_co_reports: Translation json problem
In 16 we changed the translation to be a json field. So before this pr, this
report displayed the account name as a dict with different language. This pr
correct that, now the account name is displayed as it should in the report.

When printing the fuente report, the first column was empty since the dict has
no key named "bimestre". Instead, for the fuente report, we should print the
account_name in the first column. By adding a new conditional "td", the account
name will be printed in the first column of the report.
For the ICA report, we keep it like it was by putting the bimestre in the first
column.

opw-3214033

closes odoo/enterprise#41246

X-original-commit: 230de50bafba71775d4752706dabe06e059f4dba
Signed-off-by: William André (wan) <wan@odoo.com>