[FIX] l10n_co_reports: fuente report
When printing the fuente report, the first column was empty since the dict has
no key named "bimestre". Instead, for the fuente report, we should print the
account_name in the first column. By adding a new conditional "td", the account
name will be printed in the first column of the report.
For the ICA report, we keep it like it was by putting the bimestre in the first
column.

opw-3214033

closes odoo/enterprise#40218

Signed-off-by: William André (wan) <wan@odoo.com>