[FIX] approvals: prevent attachment deletion
It is possible to delete attachments linked
to an approval request when the latter is in the approved state.

You must check that the approval request is in a state other than
'approved', 'refused' or 'cancel' to be able to delete attachments.

This creates a problem in the integrity of the information.

opw-3285655

closes odoo/enterprise#41034

X-original-commit: decdfc280448b34047a4af465e3a4bc3f1ff3243
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>