[FIX] approvals: prevent attachment deletion
It is possible to delete attachments linked
to an approval request when the latter is in the approved state.

You must check that the approval request is in a state other than
'approved', 'refused' or 'cancel' to be able to delete attachments.

This creates a problem in the integrity of the information.

opw-3285655

closes odoo/enterprise#41175

X-original-commit: 666c6c03ee242c320eedc0cb68b6ef88a4a34bb1
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>