[FIX] account_reports: GL: add Initial Balance lines for Inc/Exp accts
Currently in the General Ledger, when unfolding an Income or Expense
account, no Initial Balance gets displayed.

As such, if the report period starts in the middle of the fiscal year,
any AMLs between the start of the fiscal year and the report start date
are unaccounted for:
    - they don't appear in the unfolded account
    - they're not in the Unaffected Earnings either.

This causes a discrepancy between the total of the AMLs shown in the
unfolded account and the account total.

We fix this by ensuring that the initial balance is displayed, unless
we are at the beginning of a fiscal year, in which case it should
not be displayed.

closes odoo/enterprise#40776

Signed-off-by: John Laterre (jol) <jol@odoo.com>