[FIX] mail: no crash when scheduling activity outside of form view
Before this commit, when scheduling an activity from document
inspector, there was the following crash:

```
this.saveRecord is not a function
```

This happens because scheduling an activity necessarily reload the
chatter. Chatter is mostly in form view, but in documents the chatter
is shown in kanban. `saveRecord` does not exist, as this is a form
view-specific feature.

This commit solves the issue by not calling `saveRecord` if it
doesn't exist.

opw-3295227

closes odoo/enterprise#40775

Related: odoo/odoo#120674
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>