[IMP] sale_subscription: prevent renew or upsell if the start_date is
equal to the next_invoice_date

We don't want to allow creating child order in that case. The parent
document should be modified directly. Checking on the dates is more
efficient than the amount of invoices because free orders will see their
next_invoice_date updated but the invoice count will remain 0.

Taskid: 3274915
X-original-commit: 70ea0b07c5851d49639af78cf38141ecd0e36340
Part-of: odoo/enterprise#42338