[FIX] timesheet_grid: prevent traceback on timer stop when using any groupby
- Steps: in my timesheet, having any groupby filter active, in the list or kanban view, if you use the timer then stops a traceback appears,
this also triggers if you restart an already existing timesheet line then spot it

- issue: even tho the record you timesheeted is recorded correctly then put back into its place, the traceback shouldn't appear

- cause: The bug appears in the popRecord() method (which is triggered at every onWillUpdateProps,
but usually it just goes in the first condition and returns nothing)
So the bug comes from the `await propsList.addExistingRecord`
this func can't work when the proplist is a dynamicGroupList instead of a dynamicRecordList.

- fix: There is 2 options: could simply be fixed by finding the right group inside the the dynamicGroupList.groups,
then using it's addExisitingRecord method.
But the addExistingRecord() is triggered only when we can't find the current timerState record in the propsList
(so when we want to add the record back to the list/kanban, when we stop the timer)
and the func is supposed return it to the propsList.
But the thing is that even if i don't define the function at all it still works,
a traceback appears but the record is successfully returned to the propsList.

X-original-commit: 18d3db8e7c663650f449bd56ea91e4372e4c3475
Part-of: odoo/enterprise#42327