[FIX] account_reports: fix aged receivable amount currency
Steps to reproduce:
- Create and confirm an invoice in a foreign currency.
- Register a full payment a few days later.
- Open the Aged Receivable Report at the invoice date.

Observed behavior:
The Amout Currency column is empty.

Desired behavior:
The Amount Currency column should siplay the correct value.

The issue comes from the fact that the query run when generating
the report lines uses the `amount_residual_currency` from
`account.move.line`, which is 0 once the invoice is paid in full.

The problem is that it does not take into account the date of
the payment, which leads to an inconsistent report in this case.

So the solution is tho use the partial reconcile amounts.

A test is also added to cover this use case.

task-3282980

closes odoo/enterprise#42469

X-original-commit: 07405020917b22fa7aa1be08b4fafbaf6f805793
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>