[FIX] timesheet_grid: hide display timer button in tree view of All Timesheet
Before to this commit, the timer button was displayed in the tree view for both
My Timesheet and All Timesheet, as their tree views were the same and the
'timesheet_uom_timer' widget was utilized for the timer button in the tree view.

With this commit, the 'timesheet_uom_timer' widget has been replaced with
'timesheet_uom' in the tree view of All Timesheet.

task-3304959

X-original-commit: ed18f63a38c430de31981c74b3c8c7d5de046fff
Part-of: odoo/enterprise#42967