[FIX] l10n_es_reports: mod 303: inverted fields in BOE export
'Complementary declaration' and 'without activity' fields of the BOE export for mod 303 had been swapped, causing the file to be incorrect and get rejected when casilla 71 was 0 (because the government's site then assumed it was a complementary declaration, and expected an additional reference field to be completed).

The desired content for mod 303's file can be found here https://sede.agenciatributaria.gob.es/Sede/ayuda/disenos-registro/modelos-300-399.html

This commit keeps the same behavior as before for pre-2023 report generations. The documentation for them is not accessible anymore, so we assume the issue was introduced by the 2023 update.

OPW 3271321

closes odoo/enterprise#40832

X-original-commit: 439aebf40b1107010e292b99783398cfa8337a89
Signed-off-by: William André (wan) <wan@odoo.com>