[FIX] hr_payroll_account: salaries journal
Before this PR, when installing the module. A Traceback could happen if you already have a journal with a code SAL. This PR remove from the demo data the adding of the journal and is replaced by a post init hook to check if the journal is present, if not then add it.
In addition, we use the journal for hr_payroll_structure so in the post init hook, we had the journal for the structure that needed it.

closes odoo/enterprise#45474

Task-id: 3271099
X-original-commit: 8e62d61f5581b60e4adabf5282ee45f2e39bf603
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>