[FIX] account_accountant: don't force tax_tag_invert for write-offs in bank reconciliation widget
Partial revert of b151161 .
This fix introduced a new bug: when creating an write-off with taxes from the bank reconciliation widget, signs were all inverted in the tax report, because that commit was forcing the values of tax_tax_invert in a wrong way.

The proper fix is to fix the computation of the is_refund field in community, so that negative repartition lines are taken into account ("debit" is a refund for a "sales" tax only for positive repartition lines, for example).

Community part is in odoo/odoo#120152

X-original-commit: 91a2cfb2e33114fc8674b2f80a36bea27df646b9
Part-of: odoo/enterprise#40856