[REF] web_studio: EditorMenu's Dom, css, and api
Before this commit the editorMenu's DOM, style and API were messy.
This menu is the one, in studio, under the main navBar, that holds
the switchView buttons, the tabs and the breadcrumbs, as well as, the undo/redo
buttons if available.

After this commit, the DOM and scss have be been rewritten to take advantage of bootstrap5
as much as possible.

The undo/redo feature is now the responsibility of the component using it. That it, it should use
the hook useEditorMenuItem, to push the component in the editorMenu.

Part-of: odoo/enterprise#43441