[FIX] web_studio: show an invisible x2many field
Have a form view with an invisible x2many field (with `invisible="1"`)
In studio, edit that form view and enable the "Show invisible" switch in the sidebar.

Before this commit, that flow was not tested, leading to the feature being constantly lost and crashy.

After this commit, and in link with odoo/odoo#120918, this flow is fixed and tested.

closes odoo/enterprise#40852

Taskid: 3305808
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>