[FIX] account_asset: account_asset_id on deferred revenue
Use the account_depreciation_id as account_asset_id,
instead of account_depreciation_expense_id for
deferred revenue. This is done to avoid issue when
creating a deferred revenue, as we can not select
the deferred revenue model.

Steps:

- Take an account with type 'Current Liabilities'
- Set 'create_asset' to 'draft'
- Create a Deferred Revenue Model XXX
- Set the liability account mentioned
  as depreciation_account_id
- Set an 'Income' account as
  depreciation_account_expense_id
- Create an invoice with the liablity
  account on the invoice line, confirm
  the invoice
- Go to the revenue newly created
-> You can not select the model XXX

This is because on the revenue view, we look
for a model_id that has an 'account_asset_id'
of the same type of the one on the revenue itself,
which is, in this case, the liability account
mentioned above (because we created the
revenue based on the invoice line account).
But the 'account_asset_id' on the model was
set to be the same as the 'account_depreciation_account_id',
therefore there is no match.

opw-3186210

closes odoo/enterprise#40853

Signed-off-by: William André (wan) <wan@odoo.com>