[FIX] mrp_workorder_hr: open workorder view without HR rights
Steps to reproduce the bug:
- Connect as a user without HR rights
- Go to Manufacturing > Operations > Work Orders

Problem:
an access error is raised:
The requested operation can not be completed due to security
restrictions.
Fields: - barcode (allowed for groups 'Employees / Officer: Manage
all employees')

We do a searchRead on the "hr.employee" model to load the "barcode"
field for all employees, However, since the user "U1" does not have HR
rights, a user access error occurs.

Solution:
When loading all employees, it is unnecessary to retrieve their
barcodes. Because, When a barcode is scanned, the "onBarcodeScanned"
function is triggered, in which we call the "get_employee_barcode"
function on the Python side. This function performs a searchRead
operation as a superuser to retrieve the specific employee.

https://github.com/odoo/enterprise/blob/948358cb96cf32a153df95a8b9efa0bb56f5a53d/mrp_workorder_hr/static/src/views/hooks/employee_hooks.js#L215-L218

opw-3297355

closes odoo/enterprise#40906

X-original-commit: 506e156bc5f245f8e8397cb3599fb304d5973b15
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>