[FIX] l10n_be_hr_contract_salary: fix value on advantage change
Before this commit, emptying the value of an advantage on the
salary configurator would lead to an error because an empty
text value cannot be converted to float.

This commit fixes that issue by setting the value default to 0
for an empty field.

Steps to reproduce:
- go on an employee contract (on My Belgian Company for more
advanced advantages in demo data)
- generate a simulation link and follow the link
- select Public Transportation and add a numbered value
- unselect the field to update the simulation
- empty the field
- unselect the field to update the simulation once again

closes odoo/enterprise#40923

X-original-commit: 47ca9655b717602ae1ae9087cd6af9e78ff19adc
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>