[IMP] knowledge: force values of 'last_edition_uid' and 'last_edition_date'
It was currently possible for users with write access to the article to 'fake'
the 'last edited by' information on an article by setting those values to
someone else (and/or to another date).

Although it is not considered as a big security risk, it is better to protect
those meta-information as it does not make sense to be able to alter those
values, which are now forced at create/write to the current user.

Note that this commit does not add tests as the feature is already correctly
covered by 'test_fields_edition' in 'test_knowledge_article_internals'.

Task-3056280

Part-of: odoo/enterprise#38575