[FIX] hr_payroll: add payroll followers only for notes with payroll tag
The issue happens when adding a tag to a note, this automatically subscribes some users. The bug happens only if the payroll app is installed, because in the write method, the method note_subscribe_payroll_users is called for notes_without_tag - notes_with_tag, but it should be for the intersection of these two list, once we want to add followers for the notes that don't have the payroll tag before but have it after the update

To fix this issue, we just make sure to call it only for notes with this tag.

How to reproduce:
1. Install payroll and note apps
2. Create a note through payroll interface with payroll tag
3. Create a new note through note interface -> add a new random tag and save it
4. New followers have been automatically added to this note

closes odoo/enterprise#40892

Opw: 3286953
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>