[IMP] sale_subscription: allow cancelling subscription
When 2 subscriptions were started for the same customer by error,
it is not possible to cancel one of them, you can only close it but it
creates a churn event when it was not a churn as the customer is still there.

This PR allow cancelation of subscription that have not been invoiced
yet.

task-id : 3245825

closes odoo/enterprise#40894

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>